fix for zip of zips bug on mac and linux.  Compare this to the previous version and note that it never should have worked on any platform.
git-svn-id: https://svn.broadinstitute.org/genepattern/branches/v2_0_1@2927 48a7a3f7-ad46-0410-b969-29950918272e

+++ website/installZip.jsp
@@ -99,6 +99,7 @@ for (Iterator iter = params.iterator(); iter.hasNext();){
 		}
 		File zipFile = new File(System.getProperty("java.io.tmpdir"),name);
 		fi.write(zipFile);
+
 		requestParameters.put(fi.getFieldName(), zipFile);
 
 	}
@@ -122,6 +123,7 @@ int access_id = (request.getParameter(GPConstants.PRIVACY) != null ? GPConstants
 boolean askedRecursive = (requestParameters.get("askedRecursive") != null);
 boolean doRecursive = (requestParameters.get("inclDependents") != null);
 String fileURL = null;
+boolean deleteFile = true;
 
 if (isEncodedPost) {
 	attachedFile = (File)requestParameters.get("file1");
@@ -186,6 +188,7 @@ try {
   		              	 </body>
                 	        	</html>
 <%
+					deleteFile = false;
 					return;
 
 			}
@@ -252,7 +255,7 @@ try {
 	if (vProblems == null) vProblems = new Vector();
 	vProblems.add("Unable to install " + fullName + ": " + ioe.getMessage());
 } finally {
-	attachedFile.delete(); 
+	if (deleteFile) attachedFile.delete(); 
 }
